+++ src/main/java/org/stellar/sdk/requests/OperationsRequestBuilder.java
@@ -1,5 +1,7 @@
 package org.stellar.sdk.requests;
 
+import com.google.common.base.Joiner;
+import com.google.common.collect.Sets;
 import com.google.gson.reflect.TypeToken;
 import okhttp3.HttpUrl;
 import okhttp3.OkHttpClient;
@@ -9,6 +11,7 @@ import org.stellar.sdk.responses.Page;
 import org.stellar.sdk.responses.operations.OperationResponse;
 
 import java.io.IOException;
+import java.util.Set;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
@@ -16,8 +19,11 @@ import static com.google.common.base.Preconditions.checkNotNull;
  * Builds requests connected to operations.
  */
 public class OperationsRequestBuilder extends RequestBuilder {
+  protected Set<String> toJoin;
+
   public OperationsRequestBuilder(OkHttpClient httpClient, HttpUrl serverURI) {
     super(httpClient, serverURI, "operations");
+    toJoin = Sets.newHashSet();
   }
 
   /**
@@ -86,6 +92,29 @@ public class OperationsRequestBuilder extends RequestBuilder {
   public OperationsRequestBuilder includeFailed(boolean value) {
     uriBuilder.setQueryParameter("include_failed", String.valueOf(value));
     return this;
+  }
+
+  /**
+   * Adds a parameter defining whether to include transactions in the response. By default transaction data
+   * is not included.
+   * @param include Set to <code>true</code> to include transaction data in the operations response.
+   */
+  public OperationsRequestBuilder includeTransactions(boolean include) {
+    updateToJoin("transactions", include);
+    return this;
+  }
+
+  protected void updateToJoin(String value, boolean include) {
+    if (include) {
+      toJoin.add(value);
+    } else {
+      toJoin.remove(value);
+    }
+    if (toJoin.isEmpty()) {
+      uriBuilder.removeAllQueryParameters("join");
+    } else {
+      uriBuilder.setQueryParameter("join", Joiner.on(",").join(toJoin));
+    }
   }
 
   /**