+++ src/test/java/org/stellar/sdk/requests/PaymentsRequestBuilderTest.java
@@ -2,8 +2,6 @@ package org.stellar.sdk.requests;
 
 import okhttp3.HttpUrl;
 import org.junit.Test;
-import org.stellar.sdk.KeyPair;
-import org.stellar.sdk.Network;
 import org.stellar.sdk.Server;
 
 import static org.junit.Assert.assertEquals;
@@ -47,6 +45,36 @@ public class PaymentsRequestBuilderTest {
     HttpUrl uri = server.payments()
             .forTransaction("991534d902063b7715cd74207bef4e7bd7aa2f108f62d3eba837ce6023b2d4f3")
             .buildUri();
+    assertEquals("https://horizon-testnet.stellar.org/transactions/991534d902063b7715cd74207bef4e7bd7aa2f108f62d3eba837ce6023b2d4f3/payments", uri.toString());
+  }
+
+
+  @Test
+  public void testIncludeTransactions() {
+    Server server = new Server("https://horizon-testnet.stellar.org");
+    HttpUrl uri = server.payments()
+        .forTransaction("991534d902063b7715cd74207bef4e7bd7aa2f108f62d3eba837ce6023b2d4f3")
+        .includeTransactions(true)
+        .buildUri();
+    assertEquals("https://horizon-testnet.stellar.org/transactions/991534d902063b7715cd74207bef4e7bd7aa2f108f62d3eba837ce6023b2d4f3/payments?join=transactions", uri.toString());
+
+    uri = server.payments()
+        .forTransaction("991534d902063b7715cd74207bef4e7bd7aa2f108f62d3eba837ce6023b2d4f3")
+        .includeTransactions(true)
+        .includeTransactions(false)
+        .order(RequestBuilder.Order.ASC)
+        .includeTransactions(true)
+        .buildUri();
+    assertEquals("https://horizon-testnet.stellar.org/transactions/991534d902063b7715cd74207bef4e7bd7aa2f108f62d3eba837ce6023b2d4f3/payments?order=asc&join=transactions", uri.toString());
+  }
+
+  @Test
+  public void testExcludeTransactions() {
+    Server server = new Server("https://horizon-testnet.stellar.org");
+    HttpUrl uri = server.payments()
+        .forTransaction("991534d902063b7715cd74207bef4e7bd7aa2f108f62d3eba837ce6023b2d4f3")
+        .includeTransactions(false)
+        .buildUri();
     assertEquals("https://horizon-testnet.stellar.org/transactions/991534d902063b7715cd74207bef4e7bd7aa2f108f62d3eba837ce6023b2d4f3/payments", uri.toString());
   }
 }